# lib/api-zod/src/generated/types/attemptInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AttemptInput {
  conceptId: number;
  selectedOption: number;
  correctOption: number;
  /** @minimum 0 */
  timeSpentSeconds: number;
  /** @minimum 0 */
  switches: number;
}