# lib/api-zod/src/generated/types/attemptResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AttemptResultNextAction } from './attemptResultNextAction';
import type { AttemptResultRemediation } from './attemptResultRemediation';

export interface AttemptResult {
  correct: boolean;
  message: string;
  competencyLevel: number;
  masteryScore: number;
  nextAction: AttemptResultNextAction;
  /** @nullable */
  remediation: AttemptResultRemediation;
}