# lib/api-zod/src/generated/types/concept.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ConceptStatus } from './conceptStatus';

export interface Concept {
  id: number;
  title: string;
  description: string;
  /**
     * @minimum 1
     * @maximum 5
     */
  difficultyLevel: number;
  status: ConceptStatus;
  competencyLevel: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  prerequisiteIds: number[];
  estimatedMinutes: number;
}