# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Activity } from './activity';
import type { Concept } from './concept';
import type { Insight } from './insight';
import type { SkillNode } from './skillNode';
import type { UserProfile } from './userProfile';

export interface Dashboard {
  profile: UserProfile;
  activeConcept: Concept;
  concepts: Concept[];
  skillNodes: SkillNode[];
  insights: Insight[];
  recentActivity: Activity[];
  /**
     * @minimum 0
     * @maximum 100
     */
  courseProgress: number;
}