# lib/api-zod/src/generated/types/diagnosticResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DiagnosticResultInterestLens } from './diagnosticResultInterestLens';
import type { DiagnosticResultPreferredStyle } from './diagnosticResultPreferredStyle';

export interface DiagnosticResult {
  preferredStyle: DiagnosticResultPreferredStyle;
  interestLens: DiagnosticResultInterestLens;
  baselineCompetency: number;
  welcomeMessage: string;
}