# lib/api-zod/src/generated/types/frictionEventResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface FrictionEventResult {
  frictionScore: number;
  helperRecommended: boolean;
  helperTitle: string;
  helperBody: string;
}