# lib/api-zod/src/generated/types/lesson.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { LessonQuestion } from './lessonQuestion';
import type { LessonSource } from './lessonSource';

export interface Lesson {
  conceptId: number;
  title: string;
  eyebrow: string;
  summary: string;
  analogy: string;
  keyPoints: string[];
  codeExample: string;
  question: LessonQuestion;
  source: LessonSource;
}