# lib/api-zod/src/generated/types/profileUpdate.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ProfileUpdateInterestLens } from './profileUpdateInterestLens';
import type { ProfileUpdatePreferredStyle } from './profileUpdatePreferredStyle';

export interface ProfileUpdate {
  /** @minLength 1 */
  name?: string;
  preferredStyle?: ProfileUpdatePreferredStyle;
  interestLens?: ProfileUpdateInterestLens;
}