# lib/api-zod/src/generated/types/skillNode.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { SkillNodeStatus } from './skillNodeStatus';

export interface SkillNode {
  id: number;
  label: string;
  status: SkillNodeStatus;
  competencyLevel: number;
  x: number;
  y: number;
}