# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserProfileInterestLens } from './userProfileInterestLens';
import type { UserProfilePreferredStyle } from './userProfilePreferredStyle';

export interface UserProfile {
  id: number;
  name: string;
  email: string;
  preferredStyle: UserProfilePreferredStyle;
  interestLens: UserProfileInterestLens;
  masteryScore: number;
  streakDays: number;
  minutesLearned: number;
}